Ignore all negative tabindex values.
While there isn't really any reasonable reason to do so,
tabindex can be set to any negative value in order to remove
it from the sequential keyboard navigation.

Since it is in the spec, and it is very easy to write an adequate
selector, there seems to be no reason not to support it.