Reintroduce removal of `aria-hidden`, bump to 5.4.1
Unclear when this change happened, but the script at some point dropped the necessary removal of `aria-hidden` on show. This is still present in the "live demo" which is why it wasn't spotted sooner (the live demo / gh-pages branch seems out of sync with the main branch).

this change also includes the minified version, and bumps the version to 5.4.1